requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
adlfs==2026.2.0
    # via -r requirements.in
aiobotocore==3.3.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via
    #   aiobotocore
    #   azure-core
aioitertools==0.13.0
    # via aiobotocore
aiosignal==1.4.0
    # via aiohttp
async-timeout==5.0.1
    # via aiohttp
attrs==26.1.0
    # via aiohttp
azure-core[aio]==1.39.0
    # via
    #   adlfs
    #   azure-identity
    #   azure-storage-blob
azure-datalake-store==0.0.53
    # via adlfs
azure-identity==1.25.3
    # via adlfs
azure-storage-blob[aio]==12.28.0
    # via adlfs
botocore==1.42.70
    # via aiobotocore
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via
    #   azure-datalake-store
    #   cryptography
charset-normalizer==3.4.7
    # via requests
cryptography==43.0.3
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
deprecated==1.2.14
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.3.0
    # via adlfs
idna==3.11
    # via
    #   requests
    #   yarl
isodate==0.7.2
    # via azure-storage-blob
jmespath==0.10.0
    # via
    #   aiobotocore
    #   botocore
msal==1.35.1
    # via
    #   azure-datalake-store
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.7.1
    # via
    #   aiobotocore
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==3.0
    # via cffi
pyjwt[crypto]==2.12.1
    # via
    #   msal
    #   pyjwt
python-dateutil==2.9.0.post0
    # via
    #   aiobotocore
    #   botocore
requests==2.33.1
    # via
    #   azure-core
    #   azure-datalake-store
    #   msal
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via
    #   aiobotocore
    #   aiosignal
    #   azure-core
    #   azure-identity
    #   azure-storage-blob
    #   multidict
    #   pyjwt
urllib3==2.6.3
    # via
    #   botocore
    #   requests
wrapt==1.17.3
    # via
    #   -r requirements.in
    #   aiobotocore
    #   deprecated
yarl==1.23.0
    # via aiohttp

#requirements.txt  
#Data processing and ML  
pandas==2.1.0  
numpy==1.26.0  
scikit-learn==1.3.2  
joblib==1.3.2  
#ML pipelines and experiment tracking  
mlflow==2.11.3  
dvc==3.50.1  
pyyaml==6.0.1  
#API and deployment  
fastapi==0.109.0  
uvicorn[standard]==0.23.2  
python-dotenv==1.0.0  
#Visualization and monitoring  
matplotlib==3.8.4  
evidently==0.4.19  
prometheus-client==0.18.0 
grafana-api==1.0.3 
#Model explainability  
shap==0.42.1  
lime==0.2.0.1 
#Data Validation  
great-expectations==1.15.1  
#PostgreSQL driver 
psycopg2-binary==2.9.11 
#optional:Additional Helpers  
tqdm==4.66.1
